Refactor of code
