+++ app/src/main/java/com/androidstudy/mpesa/repo/PaymentRepository.kt
@@ -24,14 +24,7 @@ constructor() {
     fun pay(token : String, phoneNumber: String, amount: Int, description: String): DarajaLiveData<LNMResult> {
         val listener = DarajaLiveData<LNMResult>()
 
-        daraja.makePaymentRequest(
-                token,
-                phoneNumber,
-                amount.toString(),
-                AppUtils.UUID(),
-                description,
-                listener
-        )
+        daraja.makePaymentRequest(token, phoneNumber, amount.toString(), AppUtils.UUID(), description, listener)
 
         return listener
     }
