+++ daraja/src/main/java/com/androidstudy/daraja/okhttp/UnsafeOkHttpClient.java
@@ -1,6 +1,5 @@
 package com.androidstudy.daraja.okhttp;
 
-import java.security.cert.CertificateException;
 import java.security.cert.X509Certificate;
 
 import javax.net.ssl.HostnameVerifier;
